% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicles_of_amber_package.R
\docType{package}
\name{Amber-package}
\alias{Amber}
\alias{Amber-package}
\title{Amber: Roger Zelazny's Chronicle of Amber 10 book series}
\description{
Full text of the Chronicles of Amber, written by Roger Zelazny. This text is ready for text analysis of the 10 books "Nine Princes in Amber", "The Guns of Avalon", "Sign of the Unicorn", "The Hand of Oberon", "The Courts of Chaos", "The Trumps of Doom", "Blood of Amber", "Sign of Chaos", "Knight of Shadows", and "Prince of Chaos".
}
\keyword{internal}
