% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{blood_of_amber}
\alias{blood_of_amber}
\title{The text of Roger Zelazny's novel "Blood of Amber"}
\format{
A character vector with 6433 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
blood_of_amber
}
\description{
A dataset containing the text of Roger Zelazny's 1986 novel Blood of Amber.
The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
