% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{calculate_diversities}
\alias{calculate_diversities}
\title{Calculate and plot diversity indices.}
\usage{
calculate_diversities(
  object,
  group = NA,
  indices = NULL,
  save.csv = F,
  facet = NULL,
  pal = palettes,
  ...
)
}
\arguments{
\item{object}{An object of class Aldo.}

\item{group}{Metadata group on the base of which to merge samples.}

\item{indices}{Indices to calculate. Standard is set to NULL to calculate all the available indices.}

\item{save.csv}{Whether to save a .csv file with the results.}

\item{facet}{Metadata group on the base of which to split the plot in multiple subplots}

\item{...}{Other arguments.}
}
\value{
An Aldo object.
}
\description{
Calculate and plot diversity indices.
}
