% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{A function to count the total number of reads in each sample.}
\usage{
count_reads(object)
}
\arguments{
\item{object}{An object of class amberobj.}
}
\value{
A reads count and a barplot.
}
\description{
A function to count the total number of reads in each sample.
}
