% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversitiesCalc.R
\name{diversitiesCalc}
\alias{diversitiesCalc}
\title{diversitiesCalc}
\usage{
diversitiesCalc(object)
}
\arguments{
\item{object}{Meltps of an Aldo object.}
}
\value{
dataframe
}
\description{
diversitiesCalc
}
