% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.heatmap.R
\name{do.heatmap}
\alias{do.heatmap}
\title{Title}
\usage{
\method{do}{heatmap}(x, sampdata, noclust = F, taxorder, showtaxanames)
}
\arguments{
\item{x}{a matrix}

\item{sampdata}{sample data}

\item{noclust}{choose if no clustering}

\item{taxorder}{taxon order}

\item{showtaxanames}{whether to show taxa names}
}
\value{
a heatmap
}
\description{
Title
}
