% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.tree.asvs.nclust.R
\name{do.tree.asvs.nclust}
\alias{do.tree.asvs.nclust}
\title{Title}
\usage{
\method{do}{tree.asvs.nclust}(x, nclust, pal.highlight)
}
\arguments{
\item{x}{asvs matrix}

\item{nclust}{number of clusters to create}

\item{pal.highlight}{palette}
}
\value{
a tree
}
\description{
Title
}
