% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.tree.samples.R
\name{do.tree.samples}
\alias{do.tree.samples}
\title{Title}
\usage{
\method{do}{tree.samples}(x, meta, sampletip = sampletip, pal.samples, ...)
}
\arguments{
\item{x}{asvs matrix}

\item{meta}{metadata}

\item{sampletip}{metadata column to color sample tips}

\item{pal.samples}{palette to use}

\item{...}{}
}
\value{
a tree.
}
\description{
Title
}
