% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_trimmador.R
\name{el_trimmador}
\alias{el_trimmador}
\title{Function to trim and quality-filter sequences.}
\usage{
el_trimmador(
  x,
  maxEE = c(2, 3),
  fwd = "CCTACGGGNGGCWGCAG",
  rev = "GACTACNVGGGTWTCTAATCC",
  truncLen = c(280, 220),
  truncQ = 2,
  nplots = NULL,
  ncores = NULL,
  primerRemovalMethod = c("cutadapt", "dada2"),
  cutadaptPath = "/usr/bin/cutadapt"
)
}
\arguments{
\item{x}{A qualityCheck object.}

\item{maxEE}{Maxium error. Standard is c(2,3)}

\item{fwd}{Sequence of the forward primer.}

\item{rev}{Sequence of the reverse primer.}

\item{truncLen}{Length for truncation.}

\item{truncQ}{Quality threshold for truncation.}

\item{nplots}{Number of quality plots to generate. Two is enough to check if primers were removed.}

\item{ncores}{Number of cores to use for parallelization. For windows is 1.}

\item{primerRemovalMethod}{Method to use for primer removal. Can be "cutadapt" or "dada2".}

\item{cutadaptPath}{Path where cutadapt is installed. Usually bin/cutadapt.}
}
\value{
A filteredSample object.
}
\description{
Function to trim and quality-filter sequences.
}
