% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{filter_dataset}
\alias{filter_dataset}
\title{A function to count the total number of reads in each sample.}
\usage{
filter_dataset(object, sampnames = F, min_reads = F)
}
\arguments{
\item{object}{An object of class Aldo.}
}
\value{
A reads count and a barplot.
}
\description{
A function to count the total number of reads in each sample.
}
