% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxonomy.R
\name{filter_taxonomy}
\alias{filter_taxonomy}
\title{A function to count the total number of reads in each sample.}
\usage{
filter_taxonomy(object, ...)
}
\arguments{
\item{object}{An object of class Aldo.}
}
\value{
A reads count and a barplot.
}
\description{
A function to count the total number of reads in each sample.
}
