% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_and_palettes.R
\name{get.pal}
\alias{get.pal}
\title{Function to get a palette of a certain number of colors}
\usage{
get.pal(x, pal = NA)
}
\arguments{
\item{x}{object.}

\item{pal}{palette.}
}
\value{
palette.
}
\description{
Function to get a palette of a certain number of colors
}
