% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.addBWtext}
\alias{intFun.addBWtext}
\title{Add text to raster}
\usage{
intFun.addBWtext(myRaster, myDigits = 1, myCex = 0.7)
}
\arguments{
\item{myRaster}{A raster object}

\item{myDigits}{An integer that gives the number of desired digits}

\item{myDigits}{A number that determines the size of the text, e.g. 0.7}
}
\value{
plots text of a raster object
}
\description{
This function adds text to raster. A threshold value determines
whether the text is black or white
}
\examples{

library(raster)
# make some data
data <- runif(100,-1,1)
data <- matrix(data, ncol=10)
data <- raster::raster(data)

plot(data)
intFun.addBWtext(myRaster = data, myDigits = 1, myCex = 0.7)

}
\keyword{internal}
