% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.anom}
\alias{intFun.anom}
\title{Anomalies}
\usage{
intFun.anom(x)
}
\arguments{
\item{x}{An R object}
}
\value{
An R object that gives the anomalies
}
\description{
This function computes anomalies.
}
\examples{

# make some data
month <- seq(1,12,1)
month <- rep(month,10)
mod <- runif(length(month), 0,100)
mod <- data.frame(month, mod)
# compute anomalies
mod.anom <- data.frame(apply(mod[2], 2, intFun.anom))

}
\keyword{internal}
