% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.bin}
\alias{intFun.bin}
\title{Make bins}
\usage{
intFun.bin(x, bin)
}
\arguments{
\item{x}{A raster object}

\item{bin}{A number}
}
\value{
A raster object with binned values
}
\description{
This function converts values into bins.
}
\examples{

library(raster)
data <- raster::raster(matrix(runif(100,0,100), ncol=10))
my.bin <- 25
# convert values to bins
bins <- intFun.bin(data, my.bin)
# plot data
plot(data);text(data)
plot(bins); text(bins)

}
\keyword{internal}
