% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.isRaster}
\alias{intFun.isRaster}
\title{Reproject coastline}
\usage{
intFun.isRaster(x)
}
\arguments{
\item{x}{An R object such as a raster or a number}
}
\description{
This function assesses whether an R object is a raster. The original
code was copied from intFun.isRaster (spatial.tools_1.6.0)
}
\examples{

x <- 1
intFun.isRaster(x)
y <- raster::raster(matrix(seq(1,10), ncol = 2))
intFun.isRaster(y)

}
\keyword{internal}
