% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.min.max.int.diff}
\alias{intFun.min.max.int.diff}
\title{Range and interval for color bar legend (difference)}
\usage{
intFun.min.max.int.diff(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Minimum, maximum, and interval value for color bar legend used for
plotting biases.
}
\description{
This function returns the minimum, maximum, and interval value
that can be used to define a color bar legend for plotting differences. The
function ensures that the absolute values of the minimum and maximum are
identical. For instance, consider a bias that ranges from -10 to 300.
The resulting minimum and maximum value used by the color scheme are then
-300 and 300, respectively. The function is similar to intFun.min.max.int.bias,
but does not exclude extremes.
}
\examples{

library(raster)
library(classInt)
# create a raster object
data <- runif(100,-23,864)
data <- matrix(data, ncol=10)
data <- raster::raster(data)
# Get min, max, and interval for color bar legend
mmi <- intFun.min.max.int.diff(data)

}
\keyword{internal}
