% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.sampleDataFrame}
\alias{intFun.sampleDataFrame}
\title{random sampling across space and time for a data frame (FLUXNET and runoff).}
\usage{
intFun.sampleDataFrame(ref)
}
\arguments{
\item{ref}{A raster object}
}
\value{
A raster object
}
\description{
This function conducts random sampling with replacement for a single raster layer
}
\examples{

# make some data
data <- seq(1, 29, 1)
data <- c(data, NA)
data <- data.frame(matrix(data, ncol = 10))

randomSample <- intFun.sampleDataFrame(ref = data)
print(data)
print(randomSample)

}
\keyword{internal}
