% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intFun.R
\name{intFun.site.points}
\alias{intFun.site.points}
\title{Make spatial points}
\usage{
intFun.site.points(lon, lat, data)
}
\arguments{
\item{lon}{A vector (longitude)}

\item{lat}{A vector (latitude)}

\item{data}{A vector (data values)}
}
\value{
A spatial points data frame
}
\description{
This function creates a spatial points data frame for a regular gaussian grid.
}
\examples{

library(raster)
# make some data
lon <- runif(100,-180,180)
lat <- runif(100, -90, 90)
data <- runif(100, 0, 10)
# convert data to spatial points data frame
data <- intFun.site.points(lon, lat, data)
plot(data)

}
\keyword{internal}
