% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{knight_of_shadows}
\alias{knight_of_shadows}
\title{The text of Roger Zelazny's novel "Knight of Shadows"}
\format{
A character vector with 6449 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
knight_of_shadows
}
\description{
A dataset containing the text of Roger Zelazny's 1989 novel Knight of
Shadows. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
