% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeitshiny.R
\name{makeitshiny}
\alias{makeitshiny}
\title{Function to clean and prepare the dataset.}
\usage{
makeitshiny(
  x,
  kd = NULL,
  renameasvs = F,
  renamesamples = F,
  comp = F,
  rmSingletons = F,
  min_abun = F
)
}
\arguments{
\item{x}{the phyloseq object.}

\item{kd}{kingdom/domain to keep.}

\item{renameasvs}{if the function should rename ASVs (ASV1, ASV2, etc...).}

\item{comp}{if it should also create a compositional dataset.}

\item{tax_clean}{if it should clean the taxonomy. If a taxonomic level is NA it is changed to the name of taxonomic level + 1. If family is "Rhizobiaceae" and genus is NA the genus becomes "Rhizobiaceae (Family)"}
}
\value{
an object of class Aldo
}
\description{
Function to clean and prepare the dataset.
}
