% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAndTaxonomy.R
\name{mergeAndTaxonomy}
\alias{mergeAndTaxonomy}
\title{Function to merge reads and assign taxonomy.}
\usage{
mergeAndTaxonomy(
  x,
  taxAssignment = "decipher",
  taxonomy_file,
  metadataFile = NULL,
  ncores = NULL
)
}
\arguments{
\item{x}{A "filteredSamples" object.}

\item{taxAssignment}{The method to assign the taxonomy.}

\item{taxonomy_file}{The file containing the taxonomy.}

\item{metadataFile}{The file with metadata.}

\item{ncores}{Number of processors to use. For Windows is 1.}
}
\value{
A finalObject object.
}
\description{
Function to merge reads and assign taxonomy.
}
