% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{nine_princes_in_amber}
\alias{nine_princes_in_amber}
\title{The text of Roger Zelazny's novel "Nine Princes in Amber"}
\format{
A character vector with 5410 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
nine_princes_in_amber
}
\description{
A dataset containing the text of Roger Zelazny's 1970 novel Nine Princes
in Amber. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
