% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova.R
\name{plot_anova}
\alias{plot_anova}
\title{Title plot_anova}
\usage{
plot_anova(x, taxlevel, group, method, showp)
}
\arguments{
\item{x}{an amberobj}

\item{taxlevel}{taxonomic level to which agglomerate the object}

\item{group}{metadata group}

\item{method}{vegdist method}
}
\value{
a list
}
\description{
Title plot_anova
}
