% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars.R
\name{plot_bars}
\alias{plot_bars}
\title{Function to make barplots.}
\usage{
plot_bars(
  object,
  taxLevel = NULL,
  comp = TRUE,
  group = NULL,
  facet = NULL,
  facet1 = NULL,
  topx = NULL,
  fill_others = F,
  pal = NULL
)
}
\arguments{
\item{object}{An Aldo object.}

\item{taxLevel}{Taxonomic level to plot.}

\item{comp}{If to plot relative abundance or not.}

\item{group}{Metadata column for sample merging to use as x axis.}

\item{facet}{Metadata column to use to split plots.}

\item{topx}{The number of most abundant taxa to plot.}

\item{pal}{Palette}

\item{...}{Other.}
}
\value{
ggplot object
}
\description{
Function to make barplots.
}
