% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Title}
\usage{
plot_heatmap(
  x,
  taxlevel,
  ntaxa,
  sampletip,
  nclust = NULL,
  cluster = "all",
  showtaxanames = T,
  addTaxaNames = F,
  pal.highlight = "pokemon",
  pal.samples = "pokemon2"
)
}
\arguments{
\item{x}{abmerobj}

\item{taxlevel}{xx}

\item{ntaxa}{xx}

\item{sampletip}{xx}

\item{nclust}{xx}

\item{cluster}{xx}

\item{showtaxanames}{xx}

\item{pal.highlight}{xx}

\item{pal.samples}{xx}
}
\value{
xx
}
\description{
Title
}
