% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiversities.R
\name{plotDiversities}
\alias{plotDiversities}
\title{Function to plot diversity indices}
\usage{
plotDiversities(object, group = NULL, facet = NULL)
}
\arguments{
\item{object}{Object.}

\item{group}{Group for boxplot coloring.}

\item{facet}{Group for faceting.}
}
\value{
plot.
}
\description{
Function to plot diversity indices
}
