% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePrimersCutadapt.R
\name{removePrimersCutadapt}
\alias{removePrimersCutadapt}
\title{Function to remove primers using cutadapt.}
\usage{
removePrimersCutadapt(
  object,
  cutadaptPath = NULL,
  FWD = "CCTACGGGNGGCWGCAG",
  REV = "GACTACNVGGGTWTCTAATCC",
  nproc = F
)
}
\arguments{
\item{object}{A "qualityCheck" object.}

\item{cutadaptPath}{Path where cutadapt is installed. Usually bin/cutadapt.}

\item{FWD}{Sequence of the forward primer.}

\item{REV}{Sequence of the reverse primer.}

\item{nproc}{Number of processors to use for parallelization. For windows set 1.}
}
\value{
A cutadapted object.
}
\description{
Function to remove primers using cutadapt.
}
