% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_setup.R
\name{sample_setup}
\alias{sample_setup}
\title{Function to organize the Fastq files and check the quality profiles.}
\usage{
sample_setup(path, patternF, patternR, nplots)
}
\arguments{
\item{path}{Path to the folder containing the Fatsq files.}

\item{patternF}{Pattern to recognize the forward files.}

\item{patternR}{Pattern to recognize the reverse files.}

\item{nplots}{Number of quality plots for forward and reverse. Two is usually enough.}
}
\value{
A qualityCheck object.
}
\description{
Function to organize the Fastq files and check the quality profiles.
}
