% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save.graph}
\alias{save.graph}
\title{A function to save the last obtained plot. Images are saved in the "figures" subfolder of the parent directory. Each image is saved as .png, .svg, and .ps.}
\usage{
save.graph(name = NULL, w = 8, h = 8)
}
\arguments{
\item{name}{Name to use for the image.}

\item{w}{Width.}

\item{h}{Height.}
}
\value{
Nothing
}
\description{
A function to save the last obtained plot. Images are saved in the "figures" subfolder of the parent directory. Each image is saved as .png, .svg, and .ps.
}
