% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{sign_of_the_unicorn}
\alias{sign_of_the_unicorn}
\title{The text of Roger Zelazny's novel "Sign of the Unicorn"}
\format{
A character vector with 5140 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
sign_of_the_unicorn
}
\description{
A dataset containing the text of Roger Zelazny's 1975 novel Sign of the
Unicorn. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
