% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{the_courts_of_chaos}
\alias{the_courts_of_chaos}
\title{The text of Roger Zelazny's novel "The Courts of Chaos"}
\format{
A character vector with 4405 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
the_courts_of_chaos
}
\description{
A dataset containing the text of Roger Zelazny's 1978 novel The Courts of
Chaos. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
