% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{the_guns_of_avalon}
\alias{the_guns_of_avalon}
\title{The text of Roger Zelazny's novel "The Guns of Avalon"}
\format{
A character vector with 6131 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
the_guns_of_avalon
}
\description{
A dataset containing the text of Roger Zelazny's 1972 The Guns of Avalon. The
PDF of The Great Book of Amber was sourced from OceanofPDF.com and converted
to text.
}
\keyword{datasets}
