% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{the_hand_of_oberon}
\alias{the_hand_of_oberon}
\title{The text of Roger Zelazny's novel "The Hand of Oberon"}
\format{
A character vector with 5077 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
the_hand_of_oberon
}
\description{
A dataset containing the text of Roger Zelazny's 1976 novel The Hand of
Oberon. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
