% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber_novels.R
\docType{data}
\name{the_trumps_of_doom}
\alias{the_trumps_of_doom}
\title{The text of Roger Zelazny's novel "The Trumps of Doom"}
\format{
A character vector with 5441 elements
}
\source{
<https://oceanofpdf.com/authors/roger-zelazny/pdf-epub-the-great-book-of-amber-the-chronicles-of-amber-1-10-download/>
}
\usage{
the_trumps_of_doom
}
\description{
A dataset containing the text of Roger Zelazny's 1985 novel The Trumps of
Doom. The PDF of The Great Book of Amber was sourced from OceanofPDF.com
and converted to text.
}
\keyword{datasets}
